backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.91.0",
    "fastapi>=0.136.1",
    "google-genai>=1.73.1",
    "pydantic-settings>=2.14.0",
    "python-multipart>=0.0.26",
    "qdrant-client>=1.17.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.4.1",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
]
